move pipeline.RegionalWebUIInput to webui_data_adaptor_v1.RegionalInput, this avoids an import cycle is it adds more dependencies.
